# main.go
package main

import (
	"log"
	"math/rand"
	"os"
	"os/signal"
	"strconv"
	"syscall"

	"github.com/gofiber/fiber/v2"
	"github.com/gofiber/fiber/v2/middleware/logger"
	"github.com/gofiber/fiber/v2/middleware/recover"

	// swagger handler
	"github.com/gofiber/swagger"

	// docs are generated by Swag CLI, you have to import it.
	_ "fiber-docker-app/docs"
)

// Todo structure
type Todo struct {
	ID        int    `json:"id"`
	Title     string `json:"title"`
	Completed bool   `json:"completed"`
}

var todos = []Todo{
	{ID: 1, Title: "Learn Go Fiber", Completed: false},
	{ID: 2, Title: "Learn Docker", Completed: false},
}

// @title Fiber CRUD API
// @version 1.0
// @description This is a sample CRUD API with Fiber and Swagger.
// @termsOfService http://swagger.io/terms/
// @contact.name API Support
// @contact.email support@swagger.io
// @license.name Apache 2.0
// @license.url http://www.apache.org/licenses/LICENSE-2.0.html
// @host localhost:9000
// @BasePath /api/v1
func main() {
	app := fiber.New(fiber.Config{
		AppName:          "Fiber CRUD API v1.0.0",
		ServerHeader:     "Fiber",
		ReadBufferSize:   16384,
		WriteBufferSize:  16384,
		BodyLimit:        4 * 1024 * 1024,
		DisableKeepalive: false,
	})

	app.Use(logger.New())
	app.Use(recover.New())

	// Swagger route
	app.Get("/swagger/*", swagger.HandlerDefault)

	api := app.Group("/api/v1")

	// Routes
	api.Get("/todos", GetTodos)
	api.Get("/todos/:id", GetTodo)
	api.Post("/todos", CreateTodo)
	api.Put("/todos/:id", UpdateTodo)
	api.Delete("/todos/:id", DeleteTodo)

	app.Get("/ping", func(c *fiber.Ctx) error {
		return c.JSON(fiber.Map{"message": "Pong 🏓", "status": "success"})
	})

	app.Get("/health", func(c *fiber.Ctx) error {
		return c.JSON(fiber.Map{"status": "ok", "uptime": "running"})
	})

	port := os.Getenv("PORT")
	if port == "" {
		port = "9000"
	}

	quit := make(chan os.Signal, 1)
	signal.Notify(quit, os.Interrupt, syscall.SIGTERM)

	go func() {
		log.Printf("🚀 Server is starting on port %s...", port)
		log.Printf("📚 Swagger UI: http://localhost:%s/swagger/index.html", port)
		if err := app.Listen(":" + port); err != nil {
			log.Printf("❌ Error starting server: %v", err)
		}
	}()

	<-quit
	log.Println("\n🛑 Shutting down server...")

	if err := app.Shutdown(); err != nil {
		log.Printf("❌ Error during shutdown: %v", err)
	}

	log.Println("✅ Server stopped gracefully")
}

// GetTodos retrieves all todos
// @Summary Get all todos
// @Description get todos
// @Tags todos
// @Accept  json
// @Produce  json
// @Success 200 {array} Todo
// @Router /todos [get]
func GetTodos(c *fiber.Ctx) error {
	return c.JSON(todos)
}

// GetTodo retrieves a specific todo by ID
// @Summary Get a todo
// @Description get string by ID
// @Tags todos
// @Accept  json
// @Produce  json
// @Param id path int true "Todo ID"
// @Success 200 {object} Todo
// @Failure 404 {object} map[string]string
// @Router /todos/{id} [get]
func GetTodo(c *fiber.Ctx) error {
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		return c.Status(400).JSON(fiber.Map{"error": "Invalid ID format"})
	}

	for _, t := range todos {
		if t.ID == id {
			return c.JSON(t)
		}
	}
	return c.Status(404).JSON(fiber.Map{"error": "Todo not found"})
}

// CreateTodo adds a new todo
// @Summary Create a todo
// @Description add by json todo
// @Tags todos
// @Accept  json
// @Produce  json
// @Param todo body Todo true "Add todo"
// @Success 201 {object} Todo
// @Failure 400 {object} map[string]string
// @Router /todos [post]
func CreateTodo(c *fiber.Ctx) error {
	todo := new(Todo)
	if err := c.BodyParser(todo); err != nil {
		return c.Status(400).JSON(fiber.Map{"error": "Cannot parse JSON"})
	}

	todo.ID = rand.Intn(1000) + 10 // Mock ID generation
	todos = append(todos, *todo)
	return c.Status(201).JSON(todo)
}

// UpdateTodo updates an existing todo
// @Summary Update a todo
// @Description update by json todo
// @Tags todos
// @Accept  json
// @Produce  json
// @Param id path int true "Todo ID"
// @Param todo body Todo true "Update todo"
// @Success 200 {object} Todo
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /todos/{id} [put]
func UpdateTodo(c *fiber.Ctx) error {
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		return c.Status(400).JSON(fiber.Map{"error": "Invalid ID format"})
	}

	updateData := new(Todo)
	if err := c.BodyParser(updateData); err != nil {
		return c.Status(400).JSON(fiber.Map{"error": "Cannot parse JSON"})
	}

	for i, t := range todos {
		if t.ID == id {
			todos[i].Title = updateData.Title
			todos[i].Completed = updateData.Completed
			return c.JSON(todos[i])
		}
	}
	return c.Status(404).JSON(fiber.Map{"error": "Todo not found"})
}

// DeleteTodo deletes a todo by ID
// @Summary Delete a todo
// @Description delete by ID
// @Tags todos
// @Accept  json
// @Produce  json
// @Param id path int true "Todo ID"
// @Success 200 {object} map[string]string
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /todos/{id} [delete]
func DeleteTodo(c *fiber.Ctx) error {
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		return c.Status(400).JSON(fiber.Map{"error": "Invalid ID format"})
	}

	for i, t := range todos {
		if t.ID == id {
			todos = append(todos[:i], todos[i+1:]...)
			return c.JSON(fiber.Map{"message": "Todo deleted successfully"})
		}
	}
	return c.Status(404).JSON(fiber.Map{"error": "Todo not found"})
}